@@ -1,7 +1,7 @@
 {
   "name": "b2b-sellers/example-plugin",
   "description": "The example plugin shows how easy you can extend the B2Bsellers Suite (based on Shopware 6).",
-  "version": "1.0.1",
+  "version": "2.0.0",
   "type": "shopware-platform-plugin",
   "license": "MIT",
   "minimum-stability": "dev",
@@ -15,8 +15,9 @@
     }
   ],
   "require": {
-    "php": "^7.4||^8.0",
-    "shopware/core": "^6.4.0.0"
+    "php": "^8.1",
+    "shopware/core": "^6.5.6.0",
+    "b2b-sellers/core": "^2.1.0"
   },
   "require-dev": {
     "k10r/codestyle": "^3.0",
